\name{initialize-methods}
\docType{methods}

\alias{initialize,assessment-method}
\alias{initialize,geneSubsets-method}
\alias{initialize,thresholds-method}

\title{Initialize objects of class from Rmagpie}
\description{ Initialize method for Rmagpie classes.}

\author{ Camille Maumet }

\keyword{methods}