\name{show-methods}
\docType{methods}

\alias{show,assessment-method}
\alias{show,thresholds-method}
\alias{show,cvErrorRate-method}
\alias{show,cvErrorRate2ndLayer-method}
\alias{show,errorRate1stLayerCV-method}
\alias{show,errorRate2ndLayerCV-method}
\alias{show,finalClassifier-method}
\alias{show,frequencyGenes-method}
\alias{show,frequencyTopGenePerOneModel-method}
\alias{show,geneSubsets-method}
\alias{show,result2LayerCV-method}
\alias{show,resultRepeated1LayerCV-method}
\alias{show,resultRepeated2LayerCV-method}
\alias{show,resultSingle1LayerCV-method}
\alias{show,selectedGenes-method}
\alias{show,selectedGenes1stLayerCV-method}
\alias{show,selectedGenes2ndLayerCV-method}
\alias{show,selectedGenesPerOneOption-method}

\title{show Display the object, by printing, plotting or whatever suits its class}
\description{
Implementation of R method show to display object from package \code{Rmagpie}.
}

\author{ Camille Maumet }

\keyword{methods}