\name{vV70genes}
\alias{vV70genes}
\docType{data}
\title{vV70genes: van't Veer et al. 70 best genes in an object of class dataset.}
\description{
  Gene Expression values and output classes of the 70 best genes selected by van't Veer et al. (cf. references) on
78 patients in `Gene expression profiling predicts clinical outcome of breast cancer'.
}
\usage{data('vV70genesDataset')}
\format{
lksdjskh
}

\references{
 L.J. van 't Veer LJ, H. Dai, M.J. van de Vijver, Y.D. He, A.A. Hart, M. Mao,
H.L. Peterse,K. van der Kooy, M.J. Marton, A.T. Witteveen, G.J. Schreiber,
R.M. Kerkhoven, C. Roberts, P.S. Linsley, R. Bernards, S.H. Friend,
`Gene expression profiling predicts clinical outcome of breast cancer',
in Nature. 2002 Jan 31;415(6871):484-5.
}

\examples{
data('vV70genesDataset')
vV70genes
}
\keyword{datasets}
